add windows support for zadig-agent (#3246)
* add windows support for zadig-agent


* fix can't remove log/tmp files and add batch file step to aslan


* support git command on windows for zagent


* fix git and batch file step issue on windows


* support batch file for freestyle job


* add windows platform for agent access/upgrade command


* improve agent build&publish process


* update agent access/upgrade/recovery command for windows


* fix vm stream log issue


* update agent access/upgrade cmds and support shell scripts on windows


* update support shell scripts on windows


* fix bugs


* fix agent access/upgrade cmd bug


* add script type support to build template


* use getZadigAgentVersion in vm access/upgrade cmds


* fix agent archive step issue


* fix vm log issue


* fix agent report error when exec failed


---------

+++ pkg/cli/zadig-agent/internal/agent/job/job_executor.go
@@ -232,21 +232,24 @@ func (e *JobExecutor) Execute() {
 		e.ReporterCancel()
 
 		if outputs, err := e.getJobOutputVars(); err != nil {
-			e.Logger.Errorf("failed to collect job result, error: %s", err)
+			e.Logger.Errorf("failed to collect job result, error: %w", err)
 			e.JobResult.SetError(fmt.Errorf("failed to collect job result, error: %s", err))
 		} else {
 			err = e.JobResult.SetOutputs(outputs)
 			if err != nil {
-				e.Logger.Errorf("failed to set job outputs, error: %s", err)
+				e.Logger.Errorf("failed to set job outputs, error: %w", err)
 				e.JobResult.SetError(fmt.Errorf("failed to set job outputs, error: %s", err))
 			}
 		}
 
 		e.Logger.Printf("====================== Job Executor End. Duration: %.2f seconds ======================\n", time.Since(start).Seconds())
+		e.Logger.Sync()
 	}()
 	e.Logger.Printf("====================== Job Executor Start ======================\n")
 	if e.CheckZadigCancel() {
 		err = fmt.Errorf("user cancel job %s", e.Job.JobName)
+		e.Logger.Errorf(err)
+		e.JobResult.SetError(err)
 		return
 	}
 
@@ -359,9 +362,13 @@ func (e *JobExecutor) AfterExecute() error {
 	}
 
 	// -------------------------------------------- delete all temp file and dir ----------------------------------------
-	err := e.deleteTempFileAndDir()
-	if err != nil {
-		log.Errorf("failed to delete temp file and dir, error: %s", err)
+	e.Logger.Close()
+
+	if !config.GetEnableDebug() {
+		err := e.deleteTempFileAndDir()
+		if err != nil {
+			log.Errorf("failed to delete temp file and dir, error: %s", err)
+		}
 	}
 
 	return nil
